Bringing docker build up to master

+++ docker/etna-base/entrypoints/build.sh
@@ -2,9 +2,9 @@
 
 set -e
 
-if [ -n "$VERBOSE" ]; then
+#if [ -n "$VERBOSE" ]; then
   set -x
-fi
+#fi
 
 export PATH="/app/node_modules/.bin:/app/vendor/bundle/$RUBY_VERSION/bin:$PATH"
 # Default directories
@@ -23,12 +23,16 @@ if [ -z "$SKIP_RUBY_SETUP" ]; then
 #    bundle config set disable_local_branch_check true
 #    bundle config set local.etna /etna
 #  fi
-  bundle check || bundle install -j "$(nproc)"
+  # Clear any existing checkouts that occurs in development volumes.
+  rm -rf /app/vendor/bundle/*/ruby/*/bundler/gems/monoetna-*
+  bundle install -j "$(nproc)"
 
   # This jank due to the wacky way bundle enforces git repo behavior means we have to hijack the production
   # checkout.  When we can have deploys use the gem inside the container _rather than_ the external bundled one,
   # this will become simpler.
-  if [ -e ../etna ]; then
+  if [ -e ../etna ] && grep 'monoetna.git' /app/Gemfile 1>/dev/null; then
+    mkdir -p /tmp/
+    rm -rf /tmp/.git
     mv /app/vendor/bundle/*/ruby/*/bundler/gems/monoetna-*/.git /tmp/.git
     rm -rf /app/vendor/bundle/*/ruby/*/bundler/gems/monoetna-*/*
     cp -r /etna/* /app/vendor/bundle/*/ruby/*/bundler/gems/monoetna-*/
